@@ -1,38 +1,136 @@
-[project]
-name = "automation-ai"
-version = "0.1.0"
-description = "Automação de processos com IA"
-requires-python = ">=3.11"
+[build-system]
+requires = ["setuptools>=80.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
 
-dependencies = [
-  "python-dotenv",
-  "pydantic",
+# ======================================================
+# Projeto
+# ======================================================
+[project]
+name = "taxsearch"
+version = "0.1.0"
+description = "Plataforma RAG jurídico-fiscal baseada em PostgreSQL + pgvector + Gemini"
+readme = "README.md"
+requires-python = ">=3.12"
+license = { text = "Proprietary" }
+
+authors = [
+    { name = "Isaac Evangelista", email = "izavarize@gmail.com" }
 ]
 
+keywords = [
+    "rag",
+    "juridico",
+    "fiscal",
+    "postgresql",
+    "pgvector",
+    "fastapi",
+    "gemini",
+    "ai"
+]
+
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Programming Language :: Python :: 3.12",
+    "Framework :: FastAPI",
+    "Topic :: Software Development :: Libraries :: Application Frameworks",
+    "Topic :: Database :: Database Engines/Servers",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence"
+]
+
+# ======================================================
+# Dependências principais
+# (espelham o requirements.txt)
+# ======================================================
+dependencies = [
+    # API
+    "fastapi==0.128.0",
+    "uvicorn==0.40.0",
+    "python-dotenv==1.2.1",
+    "python-multipart==0.0.21",
+
+    # Config / Validation
+    "pydantic==2.12.5",
+    "pydantic-settings==2.12.0",
+
+    # Banco de dados
+    "SQLAlchemy==2.0.45",
+    "psycopg[binary]==3.3.2",
+    "pgvector==0.4.2",
+    "alembic==1.16.5",
+
+    # IA / Gemini
+    "google-genai==1.57.0",
+    "google-auth==2.47.0",
+    "httpx==0.28.1",
+    "tenacity==9.1.2",
+
+    # Embeddings / NLP
+    "sentence-transformers==5.2.0",
+    "scikit-learn==1.8.0",
+    "numpy==2.4.1",
+    "regex==2025.11.3",
+
+    # Documentos
+    "pypdf==6.6.0",
+    "python-docx==1.2.0",
+    "beautifulsoup4==4.14.3",
+    "lxml==6.0.2",
+
+    # Utilitários
+    "loguru==0.7.3",
+    "PyYAML==6.0.3",
+    "tqdm==4.67.1"
+]
+
+# ======================================================
+# Dependências opcionais
+# ======================================================
+[project.optional-dependencies]
+
+dev = [
+    "pytest==8.4.1",
+    "pytest-cov>=5.0.0",
+    "black>=24.0.0",
+    "ruff>=0.6.0",
+    "mypy>=1.10.0"
+]
+
+# ======================================================
+# URLs
+# ======================================================
+[project.urls]
+Repository = "https://github.com/izavarize/taxsearch"
+Documentation = "https://github.com/izavarize/taxsearch/docs"
+
+# ======================================================
+# Pytest
+# ======================================================
 [tool.pytest.ini_options]
+testpaths = ["tests"]
+addopts = "-ra -q"
 pythonpath = ["src"]
 
+# ======================================================
+# Black
+# ======================================================
 [tool.black]
-line-length = 88
-target-version = ["py311"]
+line-length = 100
+target-version = ["py312"]
 
+# ======================================================
+# Ruff
+# ======================================================
 [tool.ruff]
-line-length = 88
-target-version = "py311"
-fix = true
+line-length = 100
+target-version = "py312"
+select = ["E", "F", "I"]
+ignore = ["E501"]
 
-select = [
-  "E",  # pycodestyle (erros)
-  "F",  # pyflakes
-  "I",  # isort
-  "B",  # bugbear
-  "UP", # pyupgrade
-]
-
-ignore = [
-  "E501", # line length (Black resolve)
-]
-
-[tool.ruff.per-file-ignores]
-"__init__.py" = ["F401"]
-
+# ======================================================
+# Mypy
+# ======================================================
+[tool.mypy]
+python_version = "3.12"
+warn_return_any = true
+warn_unused_configs = true
+ignore_missing_imports = true
