@@ -1,77 +1,65 @@
-import uuid
-from typing import List
-
-from sqlalchemy import create_engine, text
+from collections.abc import Sequence
+from sqlalchemy import text
 from sqlalchemy.engine import Engine
 
-from app.domain.rag import DocumentChunk
+from app.domain.retrieval import RetrievedChunk
 
 
 class PostgresVectorStore:
-    def __init__(self, dsn: str) -> None:
-        self._engine: Engine = create_engine(dsn)
-
-    def upsert(
-        self,
-        chunks: List[DocumentChunk],
-        embeddings: List[List[float]],
-    ) -> None:
-        sql = text(
-            """
-            INSERT INTO vectors (id, source, content, embedding)
-            VALUES (:id, :source, :content, :embedding)
-            """
-        )
-
-        with self._engine.begin() as conn:
-            for chunk, embedding in zip(chunks, embeddings):
-                conn.execute(
-                    sql,
-                    {
-                        "id": str(uuid.uuid4()),
-                        "source": chunk.source,
-                        "content": chunk.content,
-                        "embedding": embedding,
-                    },
-                )
+    def __init__(self, engine: Engine) -> None:
+        self._engine = engine
 
     def search(
         self,
-        query_embedding: List[float],
+        query: str,
+        query_embedding: Sequence[float],
         limit: int,
         min_score: float,
-    ) -> List[DocumentChunk]:
-        """
-        Busca vetorial por similaridade de cosseno (pgvector).
-        """
-
+        alpha: float = 0.7,
+        beta: float = 0.3,
+    ) -> list[RetrievedChunk]:
         sql = text(
             """
+            WITH ranked AS (
+                SELECT
+                    source,
+                    content,
+                    1 - (embedding <=> :query_vec) AS vector_score,
+                    ts_rank_cd(
+                        search_vector,
+                        plainto_tsquery('portuguese', :query)
+                    ) AS bm25_score
+                FROM document_chunks
+            )
             SELECT
                 source,
                 content,
-                1 - (embedding <=> (:query_vec)::vector) AS score
-            FROM vectors
-            WHERE 1 - (embedding <=> (:query_vec)::vector) >= :min_score
-            ORDER BY embedding <=> (:query_vec)::vector
+                ( :alpha * vector_score + :beta * bm25_score ) AS score
+            FROM ranked
+            WHERE ( :alpha * vector_score + :beta * bm25_score ) >= :min_score
+            ORDER BY score DESC
             LIMIT :limit
             """
         )
 
-        with self._engine.begin() as conn:
+        with self._engine.connect() as conn:
             rows = conn.execute(
                 sql,
                 {
-                    "query_vec": query_embedding,
+                    "query": query,
+                    "query_vec": list(query_embedding),
                     "limit": limit,
                     "min_score": min_score,
+                    "alpha": alpha,
+                    "beta": beta,
                 },
             ).fetchall()
 
         return [
-            DocumentChunk(
-                content=row.content,
+            RetrievedChunk(
                 source=row.source,
+                content=row.content,
+                score=float(row.score),
             )
             for row in rows
         ]
